feat: rework executor
