+++ libs/@local/hashql/compiletest/Cargo.toml
@@ -33,6 +33,7 @@ rayon                     = "1.10.0"
 serde.workspace           = true
 similar-asserts.workspace = true
 snapbox                   = { version = "0.6.21", features = ["dir"] }
+termtree                  = "0.5.1"
 toml                      = "0.8.20"
 tracing                   = { workspace = true }
 tracing-subscriber        = { workspace = true, features = ["env-filter"] }
